requirements.txt
arrow==1.1.0
asgiref==3.3.4
beautifulsoup4==4.9.3
binaryornot==0.4.4
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cookiecutter==1.7.2
Django==3.2
django-anymail==8.4
django-lockdown==4.0.0
django-tailwind==2.0.1
gunicorn==20.1.0
idna==2.10
Jinja2==2.11.3
jinja2-time==0.2.0
MarkupSafe==1.1.1
mysqlclient==2.0.3
opengraphio==1.0.5
poyo==0.5.0
python-dateutil==2.8.1
python-slugify==5.0.0
pytz==2021.1
requests==2.25.1
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
text-unidecode==1.3
urllib3==1.26.4
